+++ classes/core/wpdk-ajax.php
@@ -7,15 +7,15 @@ if ( wpdk_is_ajax() ) {
    * Ajax class for extends an Ajax parent class.
    * You will use this class to extends a your own Ajax gateway class.
    *
-   *     class YouClass extends WPDKAjax {}
+   *     class YourClass extends WPDKAjax {}
    *
    * In this way you can access to `registerActions` method
    *
    * @class              WPDKAjax
    * @author             =undo= <info@wpxtre.me>
    * @copyright          Copyright (C) 2012-2013 wpXtreme Inc. All Rights Reserved.
-   * @date               2013-10-29
-   * @version            1.0.1
+   * @date               2013-10-31
+   * @version            1.0.2
    * @since              0.7.5
    *
    */
@@ -29,7 +29,7 @@ if ( wpdk_is_ajax() ) {
      * @var string $version
      */
 
-    public $version = '1.0.1';
+    public $version = '1.0.2';
 
     /**
      * Create an instance of WPXCleanFixAjax class